requirements.txt
# ── IndiaQuant MCP — Python Dependencies ──────────────────────────
# Install everything with:
#   pip install -r requirements.txt

# MCP protocol (connects server to Claude Desktop)
mcp>=1.0.0

# Indian stock market data (NSE/BSE prices, options chain, history)
yfinance>=0.2.40

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP requests (NewsAPI, Alpha Vantage)
requests>=2.31.0

# Environment variables (.env file support)
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0